requirements.txt
Flask==2.1.0
google-generativeai==0.3.1
requests==2.26.0
gunicorn==20.1.0
python-dotenv==0.19.1
Werkzeug==2.0.3